pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "cicpy"
version = "0.1.8"
authors = [
  { name="Carsten Wulff", email="carsten@wulff.no" },
]
dependencies = [
  "numpy",
  "pandas",
  "svgwrite",
  "click",
  "pyyaml",
  "matplotlib",
  "cicspi"
]
license = "MIT"
description = "Custom IC Creator Python Frontend"
readme = "README.md"
requires-python = ">=3.9"
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]

[project.scripts]
cicpy = "cicpy.cic:cli"


[project.urls]
"Homepage" = "https://github.com/wulffern/cicpy"
"Bug Tracker" = "https://github.com/wulffern/cicpy/issues"
